Some minor improvements for the build script.

+++ build.sh
@@ -9,24 +9,6 @@ ssh-keyscan github.com > .ssh/known_hosts
 
 set -x
 
-git config --global user.name DDoSolitary
-git config --global user.email DDoSolitary@gmail.com
-git config --global core.autocrlf false
-git clone https://github.com/shadowsocks/shadowsocks-libev
-git clone https://github.com/shadowsocks/simple-obfs
-git clone git@github.com:DDoSolitary/shadowsocks-libev-win -b $release_branch
-
-mkdir -p stat
-touch stat/installed.db stat/ss-commit stat/obfs-commit
-ss_commit="$(git --git-dir=shadowsocks-libev/.git show -s --format=%H)"
-obfs_commit="$(git --git-dir=simple-obfs/.git show -s --format=%H)"
-if diff /etc/setup/installed.db stat/installed.db && [ "$(cat stat/ss-commit)" == "$ss_commit" ] && [ "$(cat stat/obfs-commit)" == "$obfs_commit" ]; then
-	exit 0
-fi
-cp /etc/setup/installed.db stat/
-echo "$ss_commit" > stat/ss-commit
-echo "$obfs_commit" > stat/obfs-commit
-
 function build {
 	pushd $1
 	git submodule update --init
@@ -37,18 +19,44 @@ function build {
 	popd
 }
 
+function get-version {
+	$1 -h | awk 'NR == 2 { printf $2 }'
+}
+
+function get-commit {
+	git --git-dir="$1/.git" show -s --format=%H
+}
+
+git config --global user.name DDoSolitary
+git config --global user.email DDoSolitary@gmail.com
+git config --global core.autocrlf false
+git clone https://github.com/shadowsocks/shadowsocks-libev
+git clone https://github.com/shadowsocks/simple-obfs
+git clone git@github.com:DDoSolitary/shadowsocks-libev-win -b $release_branch
+
+mkdir -p stat
+touch stat/installed.db stat/ss-commit stat/obfs-commit
+ss_commit="$(get-commit shadowsocks-libev)"
+obfs_commit="$(get-commit simple-obfs)"
+if diff /etc/setup/installed.db stat/installed.db && [ "$(cat stat/ss-commit)" == "$ss_commit" ] && [ "$(cat stat/obfs-commit)" == "$obfs_commit" ]; then
+	exit 0
+fi
+cp /etc/setup/installed.db stat/
+echo "$ss_commit" > stat/ss-commit
+echo "$obfs_commit" > stat/obfs-commit
+
 mkdir dst
 build shadowsocks-libev
 build simple-obfs
-cp /bin/sh.exe dst/usr/local/bin
 
 cd shadowsocks-libev-win
 rm -f *.exe *.dll
-cp ../dst/usr/local/bin/* .
+cp ../dst/usr/local/bin/* /bin/sh.exe .
 cp $(ldd *.exe | awk '$3 ~ /\/usr\/bin\// { print $3 }' | sort | uniq) .
 git add -A
+
 git commit \
-	-m "Nightly build on $(date +%F) of v$(./ss-local -h | awk 'NR == 2 { print $2 }')." \
-	-m "shadowsocks/shadowsocks-libev@$(git --git-dir=../shadowsocks-libev/.git show -s --format=%H)" \
+	-m "Built on $(date +%F) for shadowsocks-libev v$(get-version ./ss-local) and simple-obfs v$(get-version ./obfs-local)." \
+	-m "shadowsocks/shadowsocks-libev@$ss_commit shadowsocks/simple-obfs@$obfs_commit" \
 	|| exit 0
 git push
